rows file 분리
