+++ .gitignore
@@ -1,0 +1,3 @@
+
+build/ios/iphonesimulator/Runner.app/Runner
+build/ios/iphonesimulator/Runner.app/Runner
